converted all GPU code to doubles

+++ src/dsp/prota/popchanfilter.cpp
@@ -32,15 +32,15 @@ using namespace boost::posix_time;
 /**************************************************************************
  * CUDA Function Prototypes
  *************************************************************************/
-extern "C" size_t gpu_channel_split(const complex<float> *h_data, complex<float> *out);
+extern "C" size_t gpu_channel_split(const complex<double> *h_data, complex<double> *out);
 extern "C" void init_deconvolve(size_t len_fft, size_t len_chan);
 extern "C" void cleanup();
 
 namespace pop
 {
 
-	PopChanFilter::PopChanFilter(): PopSink<complex<float> >( "PopChanFilter", FFT_SIZE ),
-		PopSource<complex<float> >( "PopChanFilter" ), mp_demod_func(0)
+	PopChanFilter::PopChanFilter(): PopSink<complex<double> >( "PopChanFilter", FFT_SIZE ),
+		PopSource<complex<double> >( "PopChanFilter" ), mp_demod_func(0)
 	{
 	}
 
@@ -78,14 +78,14 @@ namespace pop
 	/**
 	 * Process data.
 	 */
-	void PopChanFilter::process(const complex<float>* in, size_t len, const PopTimestamp* timestamp_data, size_t timestamp_size, size_t timestamp_buffer_correction)
+	void PopChanFilter::process(const complex<double>* in, size_t len, const PopTimestamp* timestamp_data, size_t timestamp_size, size_t timestamp_buffer_correction)
 	{
 		//size_t chan_buf_len;
 		ptime t1, t2;
 		time_duration td, tLast;
 		t1 = microsec_clock::local_time();
 
-		complex<float> *out = get_buffer(CHAN_SIZE);
+		complex<double> *out = get_buffer(CHAN_SIZE);
 
 //		//cudaProfilerStart();
 //		// call the GPU to process work
@@ -119,13 +119,13 @@ namespace pop
 
 
 		// process data
-		PopSource<complex<float> >::process(out, CHAN_SIZE, timestampOut, timestamp_size);
+		PopSource<complex<double> >::process(out, CHAN_SIZE, timestampOut, timestamp_size);
 
 		// while( chan_buf_len >= PN_SIZE)
 		// {
-		// 	complex<float> *out = get_buffer(PN_SIZE);
+		// 	complex<double> *out = get_buffer(PN_SIZE);
 		// 	chan_buf_len = gpu_demod(out);
-		// 	PopSource<complex<float> >::process();
+		// 	PopSource<complex<double> >::process();
 		// }
 		
 		
@@ -134,7 +134,7 @@ namespace pop
 		t2 = microsec_clock::local_time();
 		td = t2 - t1;
 
-		//cout << PopSource<complex<float> >::get_name() << " - 65536 RF samples received and computed in " << td.total_microseconds() << "us." << endl;
+		//cout << PopSource<complex<double> >::get_name() << " - 65536 RF samples received and computed in " << td.total_microseconds() << "us." << endl;
 	}
